+++ src/Persistence/src/Dao/SQLite/SQLiteEngineDao.h
@@ -44,7 +44,7 @@ public:
 
     virtual bool add(qint64 aircraftId, const EngineData &data) noexcept override;
     virtual bool getByAircraftId(qint64 aircraftId, QVector<EngineData> &data) const noexcept override;
-    virtual bool deleteByScenarioId(qint64 scenarioId) noexcept override;
+    virtual bool deleteByFlightId(qint64 flightId) noexcept override;
 
 private:
     std::unique_ptr<SQLiteEngineDaoPrivate> d;