+++ src/Persistence/src/Dao/SQLite/SQLiteSecondaryFlightControlDao.cpp
@@ -44,7 +44,7 @@ public:
 
     std::unique_ptr<QSqlQuery> insertQuery;
     std::unique_ptr<QSqlQuery> selectByAircraftIdQuery;
-    std::unique_ptr<QSqlQuery> deleteByScenarioIdQuery;
+    std::unique_ptr<QSqlQuery> deleteByFlightIdQuery;
 
     void initQueries()
     {
@@ -79,14 +79,14 @@ public:
 "where  sfc.aircraft_id = :aircraft_id "
 "order by sfc.timestamp asc;");
         }
-        if (deleteByScenarioIdQuery == nullptr) {
-            deleteByScenarioIdQuery = std::make_unique<QSqlQuery>();
-            deleteByScenarioIdQuery->prepare(
+        if (deleteByFlightIdQuery == nullptr) {
+            deleteByFlightIdQuery = std::make_unique<QSqlQuery>();
+            deleteByFlightIdQuery->prepare(
 "delete "
 "from   secondary_flight_control "
 "where  aircraft_id in (select a.id "
 "                       from   aircraft a"
-"                       where  a.scenario_id = :scenario_id"
+"                       where  a.flight_id = :flight_id"
 "                      );");
         }
     }
@@ -95,7 +95,7 @@ public:
     {
         insertQuery = nullptr;
         selectByAircraftIdQuery = nullptr;
-        deleteByScenarioIdQuery = nullptr;
+        deleteByFlightIdQuery = nullptr;
     }
 };
 
@@ -169,14 +169,14 @@ bool SQLiteSecondaryFlightControlDao::getByAircraftId(qint64 aircraftId, QVector
     return ok;
 }
 
-bool SQLiteSecondaryFlightControlDao::deleteByScenarioId(qint64 scenarioId) noexcept
+bool SQLiteSecondaryFlightControlDao::deleteByFlightId(qint64 flightId) noexcept
 {
     d->initQueries();
-    d->deleteByScenarioIdQuery->bindValue(":scenario_id", scenarioId);
-    bool ok = d->deleteByScenarioIdQuery->exec();
+    d->deleteByFlightIdQuery->bindValue(":flight_id", flightId);
+    bool ok = d->deleteByFlightIdQuery->exec();
 #ifdef DEBUG
     if (!ok) {
-        qDebug("SQLiteSecondaryFlightControlDao::deleteByScenarioId: SQL error: %s", qPrintable(d->deleteByScenarioIdQuery->lastError().databaseText() + " - error code: " + d->deleteByScenarioIdQuery->lastError().nativeErrorCode()));
+        qDebug("SQLiteSecondaryFlightControlDao::deleteByFlightId: SQL error: %s", qPrintable(d->deleteByFlightIdQuery->lastError().databaseText() + " - error code: " + d->deleteByFlightIdQuery->lastError().nativeErrorCode()));
     }
 #endif
     return ok;