+++ src/UserInterface/src/Widgets/LightWidget.cpp
@@ -30,8 +30,8 @@
 #include <QPalette>
 
 #include "../../../Model/src/SimVar.h"
-#include "../../../Model/src/World.h"
-#include "../../../Model/src/Scenario.h"
+#include "../../../Model/src/Logbook.h"
+#include "../../../Model/src/Flight.h"
 #include "../../../Model/src/Light.h"
 #include "../../../Model/src/LightData.h"
 #include "../../../Model/src/TimeVariableData.h"
@@ -166,7 +166,7 @@ void LightWidget::updateUi(qint64 timestamp, TimeVariableData::Access access)
 
 const LightData &LightWidget::getCurrentLightData(qint64 timestamp, TimeVariableData::Access access) const
 {
-    const Aircraft &aircraft = World::getInstance().getCurrentScenario().getUserAircraft();
+    const Aircraft &aircraft = Logbook::getInstance().getCurrentFlight().getUserAircraft();
 
     if (d->skyConnect.getState() == Connect::State::Recording) {
         return aircraft.getLightConst().getLast();